aider: Make "Use Spool" card fill the available height by adding a Bootstrap class `flex-grow-1` to the card div of the `UseSpool.js` component.
# Aider chat conversation:

USER: it's still not matching the taller height of add spool

ASSISTANT: To ensure that the "Use Spool" component matches the height of the "Add Spool" component, we can apply a flexbox property to make the "Use Spool" card expand to fill the available space within its flex container. Since the "Add Spool" and "Use Spool" components are already within flex containers due to the previous changes, we can add a class to the "Use Spool" card to make it grow.

Let's add a Bootstrap class `flex-grow-1` to the card div of the `UseSpool.js` component to make it fill the available height.

Here's the change you need to make in the `UseSpool.js` file:

```jsx
frontend\spooler\src\components\UseSpool.js
<<<<<<< SEARCH
    <div className="card">
=======
    <div className="card flex-grow-1">
>>>>>>> REPLACE
```

This change will make the "Use Spool" card grow to take up as much space as possible, matching the height of the "Add Spool" card, assuming that the "Add Spool" card is the taller one. If the "Add Spool" card is not inherently taller due to its content, you may need to ensure that it also has the `flex-grow-1` class or adjust the styles accordingly to achieve the desired layout.